dst: Synthesize responses on rejected discovery
Both the profile and endpoint discovery stacks produce errors when the
resolution cannot be handled, and we handle these errors by falling back
to an alternate stack.

This changes that behavior to synthesize responses in these cases. When
profile resolution is rejected, we create a profile response that
includes the original dst addr as a target. When resolution is rejected,
we create a resolution stream with a single endpoint (again, from the
original dst address).

This can be optimized later (i.e. to overhead of a balancer when we know
we're not talking to a replica set).